Very very bad first MVP to see if it works
I created this very very bad and simple RAG system just to see if it will work. If you want to try it out for yourself you have to create an account in Groq and get a free API key. It is free. The only downside is that it has daily limits for the number of input and output tokens.