Table extraction problem attempted fix
This is my attempt at addressing the table problem. I had some success but it is not perfect. I tried it only on ING's report.

It works by:
1) First going through all the pages of the report
2) Converts the page into an image stored in memory
3) Uses a table extractor to identify the position of tables in the page
4) If it finds one (or more) with high confidence, it gets the position of the table in the picture
5) The picture is then cropped to keep only the table
6) The picture moves from memory to a stored folder

-------- The above takes approximately 1s/page of the report --------

7) We then give the pictures to a Multimodal LLM (I used llama scout through Groq. It needs an API key but you can get one for free. Mine was free as well. Watch out for the limits though)
8) I asked ChatGPT to write a system prompt for the LLM  so that it turns the picture into CSV format
9) The CSV format is then turned into a DataFrame

Complex tables are not dealt with very well but for the majority of the tables, it produces good results. Maybe the system prompt needs a bit of tweaking but I think it is a good start.